Move auxiliar method to Util
